+++ example/mica-mqtt-client-spring-boot-example/src/main/java/net/dreamlu/iot/mqtt/client/MqttClientApplication.java
@@ -10,7 +10,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class MqttClientApplication {
 
 	/**
-	 * 先启动 mica-mqtt-client-spring-boot-example 再启动本项目，进行测试
+	 * 先启动 mica-mqtt-server-spring-boot-example 再启动本项目，进行测试
 	 */
 	public static void main(String[] args) {
 		SpringApplication.run(MqttClientApplication.class, args);