+++ starter/mica-mqtt-client-spring-boot-starter/src/main/java/net/dreamlu/iot/mqtt/spring/client/config/MqttClientConfiguration.java
@@ -59,8 +59,7 @@ public class MqttClientConfiguration {
 
 	@Bean
 	public MqttClientCreator mqttClientCreator(MqttClientProperties properties,
-											   ObjectProvider<IMqttClientSession> clientSessionObjectProvider,
-											   ObjectProvider<MqttClientCustomizer> customizers) {
+											   ObjectProvider<IMqttClientSession> clientSessionObjectProvider) {
 		MqttClientCreator clientCreator = MqttClient.create()
 			.name(properties.getName())
 			.ip(properties.getIp())
@@ -106,15 +105,14 @@ public class MqttClientConfiguration {
 		}
 		// 客户端 session
 		clientSessionObjectProvider.ifAvailable(clientCreator::clientSession);
-		// 自定义处理
-		customizers.ifAvailable((customizer) -> customizer.customize(clientCreator));
 		return clientCreator;
 	}
 
 	@Bean(MqttClientTemplate.DEFAULT_CLIENT_TEMPLATE_BEAN)
 	public MqttClientTemplate mqttClientTemplate(MqttClientCreator mqttClientCreator,
-												 ObjectProvider<IMqttClientConnectListener> clientConnectListenerObjectProvider) {
-		return new MqttClientTemplate(mqttClientCreator, clientConnectListenerObjectProvider);
+												 ObjectProvider<IMqttClientConnectListener> clientConnectListenerObjectProvider,
+												 ObjectProvider<MqttClientCustomizer> customizers) {
+		return new MqttClientTemplate(mqttClientCreator, clientConnectListenerObjectProvider, customizers);
 	}
 
 	@Bean